requirements.txt
flask
opencv-python-headless
numpy
pytesseract
deepface
python-dateutil
gunicorn
